add system property resolution to query context.
