+++ presto-spi/src/main/java/com/facebook/presto/spi/eventlistener/QueryContext.java
@@ -11,170 +11,71 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.facebook.presto.spi.eventlistener;
 
 import com.facebook.presto.spi.resourceGroups.ResourceGroupId;
 import com.facebook.presto.spi.session.ResourceEstimates;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 
-import static java.util.Objects.requireNonNull;
-
-public class QueryContext
+public interface QueryContext
 {
-    private final String user;
-    private final Optional<String> principal;
-    private final Optional<String> remoteClientAddress;
-    private final Optional<String> userAgent;
-    private final Optional<String> clientInfo;
-    private final Set<String> clientTags;
-    private final Set<String> clientCapabilities;
-    private final Optional<String> source;
-
-    private final Optional<String> catalog;
-    private final Optional<String> schema;
-
-    private final Optional<ResourceGroupId> resourceGroupId;
-
-    private final Map<String, String> sessionProperties;
-    private final ResourceEstimates resourceEstimates;
-
-    private final String serverAddress;
-    private final String serverVersion;
-    private final String environment;
-
-    public QueryContext(
-            String user,
-            Optional<String> principal,
-            Optional<String> remoteClientAddress,
-            Optional<String> userAgent,
-            Optional<String> clientInfo,
-            Set<String> clientTags,
-            Set<String> clientCapabilities,
-            Optional<String> source,
-            Optional<String> catalog,
-            Optional<String> schema,
-            Optional<ResourceGroupId> resourceGroupId,
-            Map<String, String> sessionProperties,
-            ResourceEstimates resourceEstimates,
-            String serverAddress,
-            String serverVersion,
-            String environment)
-    {
-        this.user = requireNonNull(user, "user is null");
-        this.principal = requireNonNull(principal, "principal is null");
-        this.remoteClientAddress = requireNonNull(remoteClientAddress, "remoteClientAddress is null");
-        this.userAgent = requireNonNull(userAgent, "userAgent is null");
-        this.clientInfo = requireNonNull(clientInfo, "clientInfo is null");
-        this.clientTags = requireNonNull(clientTags, "clientTags is null");
-        this.clientCapabilities = requireNonNull(clientCapabilities, "clientCapabilities is null");
-        this.source = requireNonNull(source, "source is null");
-        this.catalog = requireNonNull(catalog, "catalog is null");
-        this.schema = requireNonNull(schema, "schema is null");
-        this.resourceGroupId = requireNonNull(resourceGroupId, "resourceGroupId is null");
-        this.sessionProperties = requireNonNull(sessionProperties, "sessionProperties is null");
-        this.resourceEstimates = requireNonNull(resourceEstimates, "resourceEstimates is null");
-        this.serverAddress = requireNonNull(serverAddress, "serverAddress is null");
-        this.serverVersion = requireNonNull(serverVersion, "serverVersion is null");
-        this.environment = requireNonNull(environment, "environment is null");
-    }
+    @JsonProperty
+    public String getUser();
 
     @JsonProperty
-    public String getUser()
-    {
-        return user;
-    }
+    public Optional<String> getPrincipal();
 
     @JsonProperty
-    public Optional<String> getPrincipal()
-    {
-        return principal;
-    }
+    public Optional<String> getRemoteClientAddress();
 
     @JsonProperty
-    public Optional<String> getRemoteClientAddress()
-    {
-        return remoteClientAddress;
-    }
+    public Optional<String> getUserAgent();
 
     @JsonProperty
-    public Optional<String> getUserAgent()
-    {
-        return userAgent;
-    }
+    public Optional<String> getClientInfo();
 
     @JsonProperty
-    public Optional<String> getClientInfo()
-    {
-        return clientInfo;
-    }
+    public Set<String> getClientTags();
 
     @JsonProperty
-    public Set<String> getClientTags()
-    {
-        return clientTags;
-    }
+    public Set<String> getClientCapabilities();
 
     @JsonProperty
-    public Set<String> getClientCapabilities()
-    {
-        return clientCapabilities;
-    }
+    public Optional<String> getSource();
 
     @JsonProperty
-    public Optional<String> getSource()
-    {
-        return source;
-    }
+    public Optional<String> getCatalog();
 
     @JsonProperty
-    public Optional<String> getCatalog()
-    {
-        return catalog;
-    }
+    public Optional<String> getSchema();
 
     @JsonProperty
-    public Optional<String> getSchema()
-    {
-        return schema;
-    }
+    public Optional<ResourceGroupId> getResourceGroupId();
 
     @JsonProperty
-    public Optional<ResourceGroupId> getResourceGroupId()
-    {
-        return resourceGroupId;
-    }
+    public Map<String, String> getSessionProperties();
 
     @JsonProperty
-    public Map<String, String> getSessionProperties()
-    {
-        return sessionProperties;
-    }
+    public ResourceEstimates getResourceEstimates();
 
     @JsonProperty
-    public ResourceEstimates getResourceEstimates()
-    {
-        return resourceEstimates;
-    }
+    String getServerAddress();
 
     @JsonProperty
-    public String getServerAddress()
-    {
-        return serverAddress;
-    }
+    String getServerVersion();
 
     @JsonProperty
-    public String getServerVersion()
-    {
-        return serverVersion;
-    }
+    String getEnvironment();
 
-    @JsonProperty
-    public String getEnvironment()
+    @JsonIgnore
+    default <T> T getSystemProperty(String name, Class<T> type)
     {
-        return environment;
+        throw new UnsupportedOperationException();
     }
 }
